whiteboard separated
