+++ codejam/client/client.py
@@ -1,29 +1,27 @@
 import asyncio
-import json
 import pathlib
 import string
 import uuid
 from enum import Enum
 from random import choices, random
-from typing import Callable, Dict, Tuple, cast
+from typing import Callable, Dict, Tuple
 
 import websockets
 from kivy.app import async_runTouchApp
 from kivy.graphics import Color, Line, Rectangle
 from kivy.input import MotionEvent
 from kivy.lang.builder import Builder
-from kivy.properties import BoundedNumericProperty, ListProperty, ObjectProperty, StringProperty
-from kivy.uix.boxlayout import BoxLayout
-from kivy.uix.modalview import ModalView
+from kivy.properties import BoundedNumericProperty, ListProperty, StringProperty
 from kivy.uix.screenmanager import Screen, ScreenManager
 from kivy.uix.widget import Widget
 
+from codejam.client.widgets.whiteboard import WhiteBoard  # noqa: F401
 from codejam.server.interfaces.message import Message
 from codejam.server.interfaces.picture_message import LineData, PictureMessage, RectData
-from codejam.server.interfaces.topics import DrawOperations, ErrorOperations, Topic, TopicEnum
+from codejam.server.interfaces.topics import DrawOperations, Topic, TopicEnum
 
 root_path = pathlib.Path(__file__).parent.resolve()
-full_path = root_path.joinpath("whiteboards.kv")
+
 main_full_path = root_path.joinpath("main.kv")
 
 
@@ -177,67 +175,6 @@ class TestCanvas(Widget):
         )
 
 
-class WhiteBoard(BoxLayout):
-    """WhiteBoard"""
-
-    btn_text = StringProperty("WebSocket Connected")
-    layout = ObjectProperty(None)
-    message = StringProperty("")
-    received = StringProperty("")
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-        # TODO: refactor
-        self.draw_callbacks: Dict[str, Callable[[Message], None]] = {
-            DrawOperations.LINE.value: self.draw_line,
-            DrawOperations.RECT.value: self.draw_rectangle,
-            DrawOperations.FRAME.value: self.draw_line,
-        }
-        self.game_callbacks: Dict[str, Callable[[Message], None]] = {}
-        self.chat_callbacks: Dict[str, Callable[[Message], None]] = {}
-        self.error_callbacks: Dict[str, Callable[[Message], None]] = {
-            ErrorOperations.BROADCAST.value: self.display_error
-        }
-
-        self.callbacks: Dict[str, Dict] = {
-            TopicEnum.DRAW.value: self.draw_callbacks,
-            TopicEnum.GAME.value: self.game_callbacks,
-            TopicEnum.CHAT.value: self.chat_callbacks,
-            TopicEnum.ERROR.value: self.error_callbacks,
-        }
-
-    def on_received(self, instance: Widget, value: str) -> None:
-        """Called when received message"""
-        self.btn_text = value
-        parsed = Message(**json.loads(value))
-        callback = self.callbacks[cast(str, parsed.topic.type)][parsed.topic.operation]
-        callback(parsed)
-
-    def draw_line(self, message: Message) -> None:
-        """Draw lines from other clients"""
-        with self.canvas:
-            Color(hsv=message.value.data.colour)
-            line = Line(points=message.value.data.line, width=message.value.data.width)
-            self.ids[message.value.draw_id] = line
-
-    def draw_rectangle(self, message: Message) -> None:
-        """Draw rectangle from other clients"""
-        with self.canvas:
-            Color(hsv=message.value.data.colour)
-            rect = Rectangle(pos=message.value.data.pos, size=message.value.data.size)
-            self.ids[message.value.draw_id] = rect
-
-    @staticmethod
-    def display_error(message: Message) -> None:
-        """Display error modal."""
-        popup = ErrorPopup(
-            title=message.value.exception,
-            message=message.value.value,
-            error_code=message.value.error_id,
-        )
-        popup.open()
-
-
 class WhiteBoardScreen(Screen):
     """WhiteBoardScreen"""
 
@@ -273,26 +210,6 @@ class WhiteBoardScreen(Screen):
             print("Connection refused", e)
 
 
-class ErrorPopup(ModalView):
-    """ErrorPopup"""
-
-    title = StringProperty("")
-    message = StringProperty("")
-    error_code = StringProperty("")
-
-
-class Instructions(BoxLayout):
-    """Instructions rule"""
-
-    _canvas = ObjectProperty(None)
-
-
-class CanvasTools(BoxLayout):
-    """CanvasTools rule"""
-
-    ...
-
-
 class RootWidget(ScreenManager):
     """Root widget"""
 
@@ -300,7 +217,6 @@ class RootWidget(ScreenManager):
     game_id = StringProperty("randomGame")
 
 
-Builder.load_file(f"{full_path}")
 root_widget = Builder.load_file(f"{main_full_path}")
 
 if __name__ == "__main__":  # pragma: no cover
